ui: Add option to disable left exit in a list
Contents
Adds a flag to segdpad::List constructor that allows preventing a left press of the Dpad from calling ui_finish().

Motivation
This will force a user to select an option.
If the list houses a main menu for example, there is no need to go back.
Previously, if the enclosing Screen didn't implement a handler for List finishing, the program would freeze.